Separate essential domain logic for tasks from framework logic (went into prelude). Encapsulate everything into domain's index to keep all interfaces in a single module.
