--- domain-logic/resources/task.ts
@@ -1,2 +1,0 @@
-import { Prisma } from '@prisma/client'
-export type Task = Prisma.TaskCreateInput
